fix: var authority
