+++ src/envVariables.ts
@@ -18,10 +18,5 @@ export const envVariables: EnvVariableType = {
   VITE_OAUTH2_REDIRECT_URL: '${PEGASE_OAUTH2_REDIRECT_URL}',
   VITE_AUTHORITY: '${PEGASE_AUTHORITY}',
 };
-export const getEnvVariables = (key: keyof EnvVariableType): string => {
-  if (envVariables[key].startsWith('$')) {
-    return import.meta.env[key] as string;
-  } else {
-    return envVariables[key];
-  }
-};
+export const getEnvVariables = (key: keyof EnvVariableType): string =>
+  envVariables[key].startsWith('$') ? (import.meta.env[key] as string) : envVariables[key];
