+++ src/shared/const/authConfig.ts
@@ -17,7 +17,7 @@ interface AuthConfig {
 export const config: AuthConfig = {
   client_id: getEnvVariables('VITE_OAUTH2_CLIENT_ID'),
   redirect_uri: getEnvVariables('VITE_OAUTH2_REDIRECT_URL'),
-  authority: getEnvVariables('VITE_OAUTH2_REDIRECT_URL'),
+  authority: getEnvVariables('VITE_AUTHORITY'),
   scope: 'openid email profile',
   maxExpiresIn: 600,
 };
