+++ vite.config.ts
@@ -65,7 +65,6 @@ export default defineConfig({
     },
   },
   define: {
-    _global: {},
     APP_VERSION: JSON.stringify(process.env.APP_VERSION),
   },
   publicDir: './public',
